remove globals from web cli and organize middlewares inside router package
